Fix imgui rendering.
We should't use dont care load operation, because it discards content of
the texture. It was replaced with load operation.
Also initial state of a swapchain image is present src khr.
I am not completely sure why, but together these chages in RT info
fix the problem with black screen.